@@ -1,33 +1,50 @@
 package controller
 
 import (
+	_ "embed"
+	"encoding/json"
 	"encoding/xml"
-	"github.com/PuerkitoBio/goquery"
-	"go.uber.org/zap"
+	"log/slog"
 	"net/http"
+	"os"
 	"strconv"
 	"strings"
+
 	"tide/tide_server/db"
+
+	"github.com/PuerkitoBio/goquery"
 )
 
-var lats = map[string]float64{"abas": 44.02, "abed": 57.14, "abur": 31.58, "acaj": 13.574000000000002, "acap": 16.833, "acap2": 16.838, "acnj": 39.355, "acnj2": 39.355, "acor1": 43.364, "acya": 16.838, "adak": 51.863, "adak2": 51.863, "aden": 12.788, "agal": -10.345999999999998, "agua": 18.458, "aigi": 38.26, "ajac": 41.92, "ajac2": 41.92, "alac1": 38.338, "alac2": 38.339, "alak": 56.898, "alak2": 56.898, "alam": 37.772, "alam2": 37.772, "albo": 35.939, "albu": 37.083, "alcu": 39.835, "aler": 82.49, "alex": 31.211, "alex1": 31.212, "alge": 36.177, "alme": 36.83, "alme2": 36.832, "alofi": -19.053, "alva": 18.766, "alva2": 18.766, "amal": 18.335, "amal2": 18.335, "amas": 41.744, "ambon": -3.6830000000000003, "AN15": 43.625, "anch": 61.238, "anch2": 61.238, "ancu": -41.867, "ancu2": -41.867, "ande": 69.317, "anta": 36.836, "anto": -23.654, "anto2": -23.654, "apfl": 29.726999999999997, "apfl2": 29.726999999999997, "apla": 29.45, "apla2": 29.45, "arac": 18.48, "aracS": 18.48, "arca": 44.665, "arca2": 44.665, "aren": 38.913000000000004, "aren2": 38.913000000000004, "aric": -18.476, "aric2": -18.476, "arrc": -22.973000000000003, "arre": 28.971999999999998, "arri": 37.296, "ascen": -7.917000000000001, "ashd": 31.831, "ashd1": 31.796, "ashk": 21.85, "askl": 31.635, "asro": 12.626, "asto": 46.208, "asto2": 46.208, "atic": -16.230999999999998, "atka": 52.232, "auct": -36.830999999999996, "audi": 48.022, "audi2": 48.022, "avon": 51.5, "AZ42": 41.446999999999996, "BA05": 41.14, "ball": 54.253, "balt": -0.433, "bamd": 39.266999999999996, "bamd2": 39.266999999999996, "bame": 44.391999999999996, "bame2": 44.391999999999996, "bamf": 48.84, "bang": 54.66, "bapj": -42.891999999999996, "bara": 18.208, "barb": 17.590999999999998, "barb2": 17.590999999999998, "barc": 41.342, "barm": 52.72, "barn": -41.05, "bass": 17.29, "batr": 34.251999999999995, "bbbc": 52.16, "bbst": 32.37, "bbst2": 32.37, "bdto": 9.351, "bele": -1.442, "benc": 34.72, "benc2": 34.72, "beno": -8.767000000000001, "bgct": 41.173, "bgct2": 41.173, "bil3": 43.352, "bitu": 1.439, "blow": 18.171, "blueb": -20.444000000000003, "bmda": 32.367, "bmsa": -40.580999999999996, "bmsa2": -40.580999999999996, "bnjl": 13.443, "bodr": 37.032, "bodri": 37.031, "boma": 42.355, "bon2": 36.802, "bork": 53.556999999999995, "bouc": 43.527, "bouc2": 43.527, "boul": 50.727, "boul2": 50.727, "bour": 50.71, "bozc": 39.836, "bozy": 36.097, "bres": 48.38, "brid": 13.107999999999999, "brom": -18.000999999999998, "brpt": 21.322, "brsk": 51.401, "btny": 40.7, "btny2": 40.7, "buca": -34.639, "buca2": -34.639, "buca3": -34.639, "bull": 12.187000000000001, "bupo": -1.03, "busa": 35.093, "buve": 3.891, "buve2": 3.891, "CA02": 39.21, "cabc": 16.803, "cabo": 22.879, "cadi": 36.542, "cala": 50.968999999999994, "cala2": 50.968999999999994, "cald": -27.065, "cald2": -27.065, "call": -12.069, "calq": 13.13, "camt": 19.812, "camu": 17.887999999999998, "caph": 19.759, "carb": 36.974000000000004, "carg": 37.57, "cart": 10.390999999999998, "casc": 38.693000000000005, "cast": 51.65, "cbmd": 38.573, "cbmd2": 38.573, "ccar": 18.636, "cctx": 27.58, "cctx2": 27.58, "cdtt": 10.094, "cedr": 28.096, "ceib": 15.79, "cent": 42.967, "cent2": 42.967, "ceut": 35.9, "ceut1": 35.896, "CF06": 39.144, "chab": 25.296, "char": 43.345, "char2": 43.345, "chenn": 13.1, "cher": 49.65, "cher2": 49.65, "chia": 14.712, "chij": 27.09, "chit": -44.025, "chnr": -26.351999999999997, "chnr2": -26.351999999999997, "chrl": 11.317, "chrp": 48.863, "chrp2": 48.863, "chrs": -10.429, "chst": -41.903, "chtt": 22.333000000000002, "chuu": 7.454, "CI20": 42.093999999999994, "cila": -7.75, "cili": -7.75, "ciut": 39.988, "clst": 20.865, "cmet": -52.96, "cmet3": -52.96, "cmnj": 38.968, "cmnj2": 38.968, "cnme": 44.64, "cocb": -12.117, "coch": 9.96, "cocos": 5.556, "cois": 12.327, "colb": 6.95, "coli": -36.538000000000004, "coli2": -36.538000000000004, "colo": 6.941, "como": -11.704, "conc": 47.873999999999995, "conc2": 47.873999999999995, "cons2": -35.356, "const": -35.356, "coqu": -29.95, "coqu2": -29.95, "cor2": 43.357, "cord": 60.558, "cord2": 60.558, "corf": 39.79, "cori": 12.482999999999999, "corn": 37.945, "corr": -39.887, "corr2": -39.887, "coru": 43.367, "cove": 9.408999999999999, "cpit": -40.899, "cpla": 29.768, "cpla2": 29.768, "CR08": 39.084, "crbc": 50.04, "cres": 41.745, "cres2": 41.745, "crnl": -37.029, "crnl2": -37.029, "crom": 52.93, "csta": 44.16, "csta2": 44.148, "cstr": -42.481, "cstr2": -42.481, "CT03": 37.498000000000005, "cule": 18.301, "cule2": 18.301, "curri": 18.017, "cuxh": 53.868, "cwfl": 27.978, "cwfl2": 27.978, "cwme": 44.657, "cwme2": 44.657, "dakar": 14.675999999999998, "darw": -12.472000000000001, "davo": 7.154, "deke": 6.981, "delf": 53.327, "denh": 52.964, "denis": -3.8, "dese": -47.754, "desh": 16.305, "desi": 16.303, "dial": 30.2, "diba": 25.649, "diel": 49.55, "diel2": 49.55, "diep": 49.93, "diep2": 49.93, "djib": 11.6, "dlpr": 51.224, "dove": 51.11, "dpnc": 36.183, "dumo": -66.652, "dunk": 51.048, "dunk2": 51.048, "duqm": 19.66, "durb": -29.883000000000003, "dutc": 53.883, "dutc2": 53.883, "dzao": -12.783, "dzao2": -12.783, "east": -27.155, "east2": -27.155, "ebla": 30.05, "elak": 58.193000000000005, "elak2": 58.193000000000005, "elat": 29.502, "elpo": 9.558, "engg": -5.346, "ense": 31.848000000000003, "epme": 44.903, "epme2": 44.903, "erdek": 40.39, "erdem": 36.611, "espe": -33.871, "estr": 12.355, "euro": 51.998000000000005, "exmt": 50.617, "faja": 18.334, "fbfl": 30.671999999999997, "fbfl2": 30.671999999999997, "fer1": 43.463, "fer2": 43.476000000000006, "ferg": -19.277, "feth": 36.621, "ffcj": -10.429, "figu": 43.483999999999995, "figu2": 43.483999999999995, "fish": 52.01, "fmfl": 26.65, "fmfl2": 26.65, "fong": -8.502, "form": 38.735, "fort": -3.717, "fosm": 43.405, "fosm2": 43.405, "fpga": 32.033, "fpga2": 32.033, "fpnt": 37.806999999999995, "fpnt2": 37.806999999999995, "fren": 23.866999999999997, "frtr": 18.15, "ftfr": 14.602, "ftfr2": 14.602, "fue2": 28.493000000000002, "fuer2": 28.498, "fuka": 40.65, "futu": -14.296, "GA37": 41.21, "gamb": -23.118000000000002, "gand": 38.995, "ganm": -0.687, "garc": -7.29, "gbit": -36.189, "gcsb": -27.939, "GE25": 44.41, "geor": 19.295, "GI20": 38.784, "gibr": 36.13, "gibr2": 36.13, "gibr3": 36.13, "gij2": 43.558, "gila": 29.263, "gist": -38.675, "gokc": 40.231, "goto": 57.68, "goug": -40.349000000000004, "gpab": -28.763, "gptx": 29.31, "greg": -52.648, "greg2": -52.648, "groo": -13.86, "guam": 13.437999999999999, "gvd9": 34.848, "gwda": 25.112, "hade": 32.471, "hade2": 32.472, "haif": 32.821999999999996, "hain": 16.0, "hako": 41.78, "hali": 44.67, "hamb": 6.124, "hana": 43.28, "hani": 6.767, "harl": 53.176, "harv": 34.467, "harw": 51.95, "hasp": 50.851000000000006, "hbay": 51.382, "helg": 54.176, "hens": 54.183, "herb": 47.026, "herb2": 47.025, "heys": 54.03, "hie2": 27.784000000000002, "hien": -20.693, "hill": -31.826, "hilo": 19.733, "hilo2": 19.733, "hink": 51.21, "hirt": 57.6, "hiva": -9.805, "hmda": 34.9, "hoek": 51.977, "holy": 53.31100000000001, "holy2": 53.31, "honn": 70.983, "hono": 21.307, "hono2": 21.307, "horn": 54.758, "hrak": 35.35, "htst": 21.298000000000002, "huahi": -16.722, "huas2": -28.469, "huas3": -28.469, "huat": 15.75, "huat2": 15.75, "huel": 37.132, "iaix": 46.007, "ibiz": 38.911, "iclr": 18.534000000000002, "iera": 35.0, "igne": 41.888999999999996, "iler": 5.284, "iler2": 5.284, "ilfa": 51.21, "IM01": 43.876999999999995, "im02": 43.878, "im02a": 43.878, "imbi": -28.230999999999998, "imbt": -28.230999999999998, "immi": 53.63, "imuj": 21.255, "imuj2": 21.255, "inav": 10.181000000000001, "inav2": 10.181000000000001, "inha": -23.916999999999998, "iqui": -20.205, "iqui2": -20.205, "isab": 18.15, "ishig": 24.33, "iske": 36.594, "ista": 41.16, "IT45": 42.119, "itea": 38.43, "jaca": 18.230999999999998, "jack": -43.973, "jask": 25.63, "jask2": 25.63, "jers": 49.18, "john": 16.739, "jrmi": 18.643, "juan": -33.635999999999996, "juan2": -33.635999999999996, "kaci": 44.416000000000004, "kahu": 20.898000000000003, "kahu2": 20.898000000000003, "kait": -42.413000000000004, "kala": 37.022, "kalt": 36.11, "kant": -2.801, "kapi": 1.078, "kaps": 36.14, "kara": 24.811999999999998, "kast": 35.514, "kata": 37.64, "kaum": 20.78, "kawa": 20.035999999999998, "kawa2": 20.035999999999998, "keba": 24.13, "kepo": -54.28, "kepo1": -54.28, "kerg": -49.345, "kerg2": -49.352, "khol": 47.051, "kiel": 54.5, "kinl": 58.46, "kis1": 53.312, "kjni": -29.059, "kodi": 57.732, "kodi2": 57.732, "koli": -6.107, "koli2": -6.107, "koro": 36.8, "kors": 46.633, "kos1": 36.898, "kos2": 36.891, "kota": 7.817, "kril": 45.9, "krna": 6.197, "kuri": 45.23, "kush": 42.98, "kusm": 33.48, "kwaj": 8.737, "kwaj2": 8.737, "kwfl": 24.553, "kwfl2": 24.553, "kypa": 37.26, "LA23": 35.5, "LA38": 44.097, "lago": 28.088, "lagos": 6.421, "lajo": 32.867, "lajo2": 32.867, "lalb": 13.485, "lali": -2.218, "lame": 18.317999999999998, "lame2": 18.317999999999998, "lamu": -2.267, "lang": 43.347, "lank": 6.416, "lapa": 28.678, "larn": 34.916, "larp": 46.158, "larp2": 46.158, "laru2": -4.5, "lasp": 28.141, "lata": -10.720999999999998, "laun": 13.309000000000001, "laza": 17.94, "lcst": 38.908, "lebu": -37.594, "lebu2": -37.594, "leco": 48.36, "leco2": 48.36, "lecy": 47.543, "lega": 13.145999999999999, "leha": 49.482, "leha2": 49.482, "leit": 55.99, "lemba": -8.731, "lena": -19.533, "lero": 14.677, "lerw": 60.155, "lerw2": 60.155, "leso": 46.498000000000005, "leso2": 46.498000000000005, "levu": -17.605, "lgos": 37.099000000000004, "LI11": 43.54600000000001, "lifo": -20.918000000000003, "lime": 17.695, "lime2": 17.695, "limon": 9.988999999999999, "lirf": -17.133, "litz": -16.113, "live": 53.45, "llan": 53.33, "lmma": 39.872, "lomb": -2.042, "losa": 33.719, "losa2": 33.719, "losu": 9.65, "lott": -37.55, "lowe": 52.47, "lpaz": 24.267, "lpaz2": 24.267, "lpbc": 54.2, "lton": 25.81, "luba": 13.818, "luga": -15.515999999999998, "lymg": 50.74, "made": 14.712, "madry": -42.763000000000005, "maer": 40.766999999999996, "magi": 17.97, "magi2": 17.97, "maho": 39.893, "mais": -46.867, "mais2": -46.867, "maji": 24.518, "maju": -6.189, "maka": 21.32, "make": -16.627, "mal3": 36.711999999999996, "mala": 7.327999999999999, "male": 4.19, "malh": 55.372, "mali": 55.372, "malk": 43.87, "malo": 61.933, "malp": 4.002, "malp2": 4.002, "malt": 35.986, "mang": 43.801, "mani": 14.582, "manz": 19.064, "mare": -21.548000000000002, "mari": 42.406000000000006, "marm": 15.41, "mars": 43.291000000000004, "marsh": 7.106, "masi": 20.683000000000003, "mata": -17.000999999999998, "matel": 38.928000000000004, "maya": 18.218, "maza": 23.180999999999997, "maza2": 23.180999999999997, "MC41": 42.743, "ME13": 38.196, "meji": -23.098000000000003, "meji2": -23.098000000000003, "meli": 35.291, "ment": 38.428000000000004, "mera": 34.92, "meul": 4.127, "meul2": 4.127, "mhav": 51.71, "midx": 28.206999999999997, "midx2": 28.206999999999997, "mill": 55.75, "mimz": 44.211000000000006, "mimz2": 44.211000000000006, "mind": 16.887999999999998, "mini": 8.28, "mins": 24.28, "mnkt": -37.047, "mnza": 19.061, "moku": 21.436999999999998, "momb": -4.067, "mona": 18.09, "mona2": 18.09, "monc": 43.729, "monc2": 43.729, "mont": 36.605, "mont2": 36.605, "mony": 41.048, "mony2": 41.048, "motr": 36.72, "moul": 16.483, "mrig": 15.548, "mrms": 36.838, "ms001": 6.257000000000001, "ms002": 5.4510000000000005, "ms003": 5.684, "ms004": 5.933, "ms005": 6.877999999999999, "ms006": 5.0760000000000005, "mtwa": -10.267000000000001, "mumb": 51.57, "murc2": 37.596, "mus2": 23.628, "musc": 23.633000000000003, "NA23": 40.84, "naga": 32.74, "naha": 26.21, "naho": 42.78, "nain": 56.55, "nanc": 8.05, "nans": 9.5, "napt": -39.476, "nauu": -0.532, "nawi": 21.956999999999997, "nawi2": 21.956999999999997, "ncpt": -34.41, "neve": 46.68, "newl": 50.103, "newl2": 50.103, "nhav": 50.78, "nhte": -38.742, "nice": 43.695, "nice2": 43.695, "niue": -19.052, "nkfa": -21.136999999999997, "nkfa2": -21.138, "nksk": 55.2, "noct": 18.1, "nome": 64.5, "nome2": 64.5, "noto": 37.5, "noua": 20.813000000000002, "npor": 51.55, "nshi": 55.01, "nspi": 40.766999999999996, "nspi2": 40.766999999999996, "ntue": -38.742, "ntue2": -38.742, "nuku": -8.915, "numbo": -22.247, "nuuk": 64.171, "nyal": 78.933, "ofun": 39.02, "ohig": -63.321000000000005, "ohig3": -63.321000000000005, "omae": 34.61, "OR24": 42.356, "oran": 12.517000000000001, "orma": 25.204, "oste": 51.233000000000004, "OT15": 40.146, "otat": -45.81399999999999, "ouin": -21.983, "ouis": 49.278999999999996, "ouis2": 49.278999999999996, "ouve": -20.55, "PA07": 38.121, "pada": -0.95, "padn": -0.95, "pagb": 13.427999999999999, "pagi": -45.162, "pagi2": -45.162, "pago": -14.277000000000001, "pago2": -14.277000000000001, "pagx": -14.277000000000001, "palb": 49.226000000000006, "pale": 35.22, "palm": 39.56, "palm1": 16.755, "pana": -7.747999999999999, "pand": -7.747999999999999, "pano": 38.36, "pant": 36.835, "papa": 10.642000000000001, "pape": -17.533, "pape2": -17.532, "papho": 34.755, "papo": -25.009, "papo2": -25.009, "para": 35.038000000000004, "parh": 17.15, "pata": -20.803, "pata2": -20.803, "pbil": 14.019, "pblu": 11.998, "pcha": -45.467, "pcha2": -45.467, "pcoi": 12.327, "pcor": 15.843, "pcrt": 12.482999999999999, "pdas": 37.73, "PE09": 37.286, "pedn": -49.13, "pedn2": -49.13, "peir": 37.935, "pemba": -12.966, "penr": -9.001, "petr": 53.016999999999996, "phbc": 50.717, "phcp": -20.04, "pich": -32.135999999999996, "pich2": -32.135999999999996, "pich3": -32.135999999999996, "pisa": -19.597, "pisa2": -19.597, "pkem": -34.474000000000004, "PL14": 40.03, "plat": -38.05, "plat2": -38.0, "plmy": 5.89, "plom": 38.972, "pluz": 28.13, "plym": 50.37, "pmel": -43.898, "pmel2": -43.898, "pmon": -41.485, "pmon2": -41.485, "pmur": -21.816999999999997, "pnfl": 30.403000000000002, "pnfl2": 30.403000000000002, "pntn": -4.783, "PO40": 40.895, "pobe": 17.474, "porl": -38.343, "poro": 49.22, "porp": 54.84, "pors": -35.109, "port": 45.568999999999996, "port2": 45.568999999999996, "posa": 33.06, "posi": 42.65, "pots": 13.01, "ppcp": 36.669000000000004, "prai": 14.912, "prat": -62.479, "prat3": -62.479, "prba": 15.695, "prec": 14.808, "preo": 42.909, "prev": 38.949, "prgi": -8.283, "pric": 12.005, "prig": -8.283, "prin": 54.32, "prin2": 54.317, "prog": 21.303, "prog2": 21.303, "prte": -33.967, "prud": 70.388, "prud2": 70.388, "prus": 55.21, "psan": 12.200999999999999, "psdn": 12.200999999999999, "psjs": 11.251, "psla": 28.932, "psla2": 28.932, "pslu": 35.168, "pslu2": 35.168, "PT17": 40.842, "ptal": -30.291999999999998, "ptal2": -30.291999999999998, "ptan": 15.665, "ptan2": 15.667, "ptar": -53.123999999999995, "ptar2": -53.123999999999995, "ptbl": 11.68, "ptca": 18.505, "ptch": -29.246, "ptfe": 43.358999999999995, "ptfe2": 43.358999999999995, "ptga": 10.138, "ptis": 50.593999999999994, "ptln": 43.00899999999999, "ptln2": 43.00899999999999, "ptlu": -20.157, "ptma": 35.921, "ptmd": 15.559000000000001, "ptmd2": 15.577, "ptme": 43.657, "ptmt": 50.8, "ptos": -0.061, "ptpl": 19.799, "ptpr": 18.534000000000002, "ptpt": 16.224, "ptpt2": 16.224, "ptro": 17.935, "ptsc": 13.263, "ptsp": 10.65, "ptve": 42.52, "ptve2": 42.52, "puer2": 20.658, "puert": 20.658, "pumo": 20.868000000000002, "pumo2": 20.868000000000002, "puyt": -46.085, "pwil": -54.933, "pwil2": -54.933, "qaqo": 60.718, "qcbc": 53.25, "qing": 19.57, "qtro": -32.775, "qtro2": -32.775, "qtro3": -32.775, "quar": 22.291, "quel": -39.398, "quel2": -39.398, "quepo": 9.425, "quin": 13.775, "quir": -36.635999999999996, "quir2": -36.635999999999996, "qura": 23.26, "RA10": 44.492, "rangi": -14.946, "raro": -21.2, "raro2": -21.205, "rbct": -29.28, "RC09": 38.122, "reun": -20.92, "reun2": -20.92, "reyk": 64.15, "rfrt": -29.250999999999998, "riki": -23.122, "rodr": -19.68, "rorv": 64.867, "rosc": 48.716, "rosc2": 48.716, "rose": 15.314, "ross": -23.160999999999998, "rote": -10.708, "rothe": -67.567, "rous": 42.636, "rous2": 42.636, "rtas": 16.346, "rudn": 44.35, "SA16": 40.677, "sab2": 5.832999999999999, "saba": 5.832999999999999, "sade": -8.19, "sado": 38.32, "sadp": -8.19, "sagr": 37.01, "sagu": 39.634, "said": 35.111999999999995, "saig": 36.2, "sain": 47.26, "sain2": 47.26, "saip": 15.227, "sala": 16.935, "sali": 16.168, "sali2": 16.168, "salv": -12.967, "Salv2": -12.967, "sama": 11.235, "sams": 41.295, "san2": 43.461000000000006, "sana": 12.569, "sand": 32.713, "sanf": -26.291999999999998, "sanf3": -26.291999999999998, "sanj": 18.462, "sanj2": 18.462, "sano": -33.582, "sano2": -33.582, "sanp": -47.70399999999999, "sant": -0.752, "saot": 0.348, "sapz": 8.66, "sapz2": 8.66, "sass": 54.511, "saum": -7.983, "SB36": 42.955, "sbea": 44.625, "sbea2": 44.625, "SC43": 37.505, "scar": 11.167, "sche": 52.099, "scoa": 43.395, "scoa2": 43.395, "scor": 70.484, "sdom": 18.421, "sdpr": 50.651, "sdpt": 55.336999999999996, "sdpt2": 55.336999999999996, "sebe": -5.936, "sebla": -3.2239999999999998, "sema": -6.9479999999999995, "sete": 43.398, "sete2": 43.398, "setp1": 26.683000000000003, "setu": 38.49, "sev2": 37.318000000000005, "seve": 50.678000000000004, "seve2": 50.678000000000004, "sewa": 60.119, "sewa2": 60.119, "shee": 51.45, "shek": 22.22, "shen": 22.47, "sho2": -33.028, "sian": 19.313, "sib2": 1.733, "sibo": 1.733, "sile": 41.18, "sino": 42.016999999999996, "sisa": 21.162, "sitin": -34.183, "sitk": 57.052, "sitk2": 57.052, "sitt": 20.15, "sjds": 11.251, "slds": 18.227, "smag": 18.297, "smag2": 18.297, "smar": -20.893, "smar2": -20.893, "smog": 58.35, "sobr": 42.793, "sole": 41.836000000000006, "sole2": 41.836000000000006, "solo": -9.429, "sona": 4.006, "sosu": 46.53, "sove": 48.97, "spjy": -43.433, "spmi": 46.778999999999996, "sprg": -42.54600000000001, "sscr": 51.57, "ST44": 38.817, "stan": -51.75, "star": 47.41, "stari": 43.181000000000004, "stcr": 17.747, "stcr2": 17.747, "sthl": -15.917, "sthl2": -15.918, "sthm": 59.33, "stjo": 47.57, "stlo2": 15.991, "stlu": 14.015999999999998, "stlu2": 13.911, "stlu3": 13.854000000000001, "stlu4": 13.720999999999998, "stma": 48.641999999999996, "stma2": 48.641999999999996, "stmr": 49.918, "stmt": 18.083, "stor": 58.21, "stor2": 58.21, "stpa": -38.711999999999996, "stqy": 48.648, "stqy2": 48.648, "subi": 14.817, "sumt": -43.57, "sura": -7.2, "suro": 22.57, "swpr": 50.608999999999995, "syow": -69.008, "syro": 37.438, "TA18": 40.476, "tako": 4.885, "tala": -4.578, "talc": -36.701, "talc2": -36.701, "talt": -25.406999999999996, "talt2": -25.406999999999996, "talt3": -25.406999999999996, "tanjo": 1.262, "tara": 1.3630000000000002, "tare": -6.693, "tari": 36.006, "tarr": 41.07899999999999, "tasu": 36.281, "taut": -37.641, "tbwc": -37.1, "tdcu": -37.05, "tela": 15.784, "telc": 21.34, "telu": 0.5539999999999999, "tema": 5.635, "tene": 28.476999999999997, "ters": 53.428000000000004, "tfbc": 49.15, "thes": 40.63, "thev": -32.149, "thio": -21.614, "thul": 76.543, "tjls": -6.483, "tkdi": 4.882, "tn031": 28.046999999999997, "tnbl": -0.53, "tnpr": 50.544, "toam2": -18.154, "tobe": 56.62, "toco2": -22.094, "toco3": -22.094, "tofi": 49.15, "tort": 18.425, "tosa": 32.78, "toul": 43.117, "toul2": 43.117, "toya": 36.76, "TR22": 45.653999999999996, "trab": 41.001999999999995, "treg": 58.0, "trin": 8.564, "trnz": 51.336000000000006, "trst": -10.585999999999999, "tubua": -23.342, "tuca": 21.434, "tudy": 47.644, "tumc": 1.82, "tumc2": 1.82, "tuxp": 20.964000000000002, "tuxp2": 20.964000000000002, "ugle": 49.08, "ulla": 57.9, "upol": -13.827, "ushu": -54.817, "util": 16.096, "vair": -17.805999999999997, "vale": 39.442, "valp": -33.028, "valp2": -33.028, "valp3": -33.028, "vanu": -17.755, "vard": 70.333, "vati": -17.398, "VE19": 45.418, "vela": 42.961999999999996, "ver1": -65.25, "vera": 19.192, "vera2": 19.192, "verav": 20.9, "vern": -65.25, "vhbc": 49.283, "VI12": 41.888000000000005, "vibc": 48.42, "vieq": 18.094, "vieq2": 18.094, "vig2": 42.243, "vil2": 42.601000000000006, "vish": 17.68, "viti": -18.134, "vlad": 43.13, "vlis": 51.443000000000005, "vodo": 51.727, "vung": 10.34, "waik": -9.39, "wait": -43.946000000000005, "waka": 45.41, "wake": 19.29, "wake2": 19.29, "wall": -13.17, "walvi": -22.949, "warn": 54.17, "wbhb": 50.71, "weym": 50.61, "whbc": 50.7, "whit": 54.49, "wick": 58.44, "will": 12.104000000000001, "wint": 50.51, "wlgt": -41.285, "wlms": 30.326, "wlms2": 30.326, "wood": 41.523, "wood2": 41.523, "work": 54.65, "wpwa": 46.908, "wpwa2": 46.908, "wuda": 23.82, "xish": 16.83, "xmas": 1.984, "yabu": 18.055, "yaku": 59.548, "yaku2": 59.548, "yalo": 40.662, "yapi": 9.514, "yuzh": 44.04, "zanz": -6.15, "zhap": 21.58, "zihu": 17.637, "zihu2": 17.636, "zygi": 34.727, "zygi1": 34.726}
-var lons = map[string]float64{"abas": 144.29, "abed": -2.08, "abur": 131.41, "acaj": -89.838, "acap": -99.917, "acap2": -99.90299999999999, "acnj": -74.418, "acnj2": -74.418, "acor1": -8.399, "acya": -99.90299999999999, "adak": -176.632, "adak2": -176.632, "aden": 44.974, "agal": 56.586000000000006, "agua": -67.165, "aigi": 22.08, "ajac": 8.76, "ajac2": 8.76, "alac1": -0.478, "alac2": -0.48100000000000004, "alak": -154.248, "alak2": -154.248, "alam": -122.29799999999999, "alam2": -122.29799999999999, "albo": -3.034, "albu": -8.26, "alcu": 3.139, "aler": -62.32, "alex": 29.924, "alex1": 29.885, "alge": -5.398, "alme": -2.478, "alme2": -2.485, "alofi": -169.921, "alva": -95.756, "alva2": -95.756, "amal": -64.92, "amal2": -64.92, "amas": 32.391999999999996, "ambon": 128.183, "AN15": 13.507, "anch": -149.888, "anch2": -149.888, "ancu": -73.833, "ancu2": -73.833, "ande": 16.15, "anta": 30.613000000000003, "anto": -70.405, "anto2": -70.405, "apfl": -84.98200000000001, "apfl2": -84.98200000000001, "apla": -91.338, "apla2": -91.338, "arac": -66.702, "aracS": -66.702, "arca": -1.1640000000000001, "arca2": -1.1640000000000001, "aren": -123.705, "aren2": -123.705, "aric": -70.32300000000001, "aric2": -70.32300000000001, "arrc": -42.013999999999996, "arre": -13.53, "arri": -8.869, "ascen": -14.417, "ashd": 34.641999999999996, "ashd1": 34.626, "ashk": 59.57, "askl": 34.494, "asro": -87.34200000000001, "asto": -123.76700000000001, "asto2": -123.76700000000001, "atic": -73.694, "atka": -174.173, "auct": 174.78599999999997, "audi": -4.537, "audi2": -4.537, "avon": -2.73, "AZ42": 12.635, "BA05": 16.866, "ball": -9.89, "balt": -90.28299999999999, "bamd": -76.578, "bamd2": -76.578, "bame": -68.205, "bame2": -68.205, "bamf": -125.14, "bang": -5.67, "bapj": 147.338, "bara": -71.092, "barb": -61.821000000000005, "barb2": -61.821000000000005, "barc": 2.166, "barm": -4.05, "barn": 145.915, "bass": -62.71, "batr": 35.656, "bbbc": -128.14, "bbst": -64.695, "bbst2": -64.695, "bdto": -82.258, "bele": -48.496, "benc": -76.67, "benc2": -76.67, "beno": 115.21700000000001, "bgct": -73.182, "bgct2": -73.182, "bil3": -3.045, "bitu": 125.19, "blow": -63.093, "blueb": 57.711000000000006, "bmda": -64.7, "bmsa": -73.737, "bmsa2": -73.737, "bnjl": -16.573, "bodr": 27.423000000000002, "bodri": 27.425, "boma": -71.05199999999999, "bon2": -6.337999999999999, "bork": 6.749, "bouc": -1.515, "bouc2": -1.515, "boul": 1.577, "boul2": 1.577, "bour": -1.87, "bozc": 26.076, "bozy": 32.941, "bres": -4.5, "brid": -59.629, "brom": 122.219, "brpt": -158.119, "brsk": 3.548, "btny": -74.02, "btny2": -74.02, "buca": -72.046, "buca2": -72.046, "buca3": -72.046, "bull": -69.02, "bupo": 100.396, "busa": 129.037, "buve": -77.081, "buve2": -77.081, "CA02": 9.113999999999999, "cabc": -88.08200000000001, "cabo": -109.91, "cadi": -6.281000000000001, "cala": 1.868, "cala2": 1.868, "cald": -70.825, "cald2": -70.825, "call": -77.167, "calq": -61.196000000000005, "camt": -90.595, "camu": -66.528, "caph": -72.193, "carb": -1.9, "carg": -0.98, "cart": -75.53699999999999, "casc": -9.415, "cast": -9.902999999999999, "cbmd": -76.068, "cbmd2": -76.068, "ccar": -91.836, "cctx": -97.21700000000001, "cctx2": -97.21700000000001, "cdtt": -61.865, "cedr": -115.18299999999999, "ceib": -86.76, "cent": 9.35, "cent2": 9.35, "ceut": -5.317, "ceut1": -5.312, "CF06": 8.308, "chab": 60.603, "char": -124.322, "char2": -124.322, "chenn": 80.3, "cher": -1.63, "cher2": -1.63, "chia": -92.40100000000001, "chij": 142.19, "chit": -176.36900000000003, "chnr": -70.634, "chnr2": -70.634, "chrl": -60.55, "chrp": -122.758, "chrp2": -122.758, "chrs": 105.669, "chst": 171.43400000000003, "chtt": 91.825, "chuu": 151.898, "CI20": 11.79, "cila": 109.0, "cili": 109.0, "ciut": 3.8280000000000003, "clst": -90.405, "cmet": -74.066, "cmet3": -74.066, "cmnj": -74.96, "cmnj2": -74.96, "cnme": -67.3, "cocb": 96.89200000000001, "coch": 76.26, "cocos": -87.04799999999999, "cois": -83.068, "colb": 79.85, "coli": -72.957, "coli2": -72.957, "colo": 79.85, "como": 43.248000000000005, "conc": -3.907, "conc2": -3.907, "cons2": -72.458, "const": -72.458, "coqu": -71.335, "coqu2": -71.335, "cor2": -8.389, "cord": -145.753, "cord2": -145.753, "corf": 19.91, "cori": -87.16799999999999, "corn": 22.936, "corr": -73.42699999999999, "corr2": -73.42699999999999, "coru": -8.4, "cove": -76.205, "cpit": 176.232, "cpla": -93.34299999999999, "cpla2": -93.34299999999999, "CR08": 17.137, "crbc": -125.25, "cres": -124.18299999999999, "cres2": -124.18299999999999, "crnl": -73.152, "crnl2": -73.152, "crom": 1.3, "csta": 28.6, "csta2": 28.671999999999997, "cstr": -73.758, "cstr2": -73.758, "CT03": 15.094000000000001, "cule": -65.303, "cule2": -65.303, "curri": 120.48299999999999, "cuxh": 8.717, "cwfl": -82.83200000000001, "cwfl2": -82.83200000000001, "cwme": -67.21, "cwme2": -67.21, "dakar": -17.42, "darw": 130.846, "davo": 125.663, "deke": 158.2, "delf": 6.933, "denh": 4.743, "denis": 55.667, "dese": -65.915, "desh": -61.79600000000001, "desi": -61.071999999999996, "dial": -88.1, "diba": 56.269, "diel": -1.86, "diel2": -1.86, "diep": 1.085, "diep2": 1.085, "djib": 43.167, "dlpr": 1.409, "dove": 1.32, "dpnc": -75.747, "dumo": 140.002, "dunk": 2.367, "dunk2": 2.367, "duqm": 57.72, "durb": 31.033, "dutc": -166.533, "dutc2": -166.533, "dzao": 45.258, "dzao2": 45.258, "east": -109.439, "east2": -109.439, "ebla": -90.368, "elak": -136.343, "elak2": -136.343, "elat": 34.918, "elpo": -78.949, "engg": 102.27799999999999, "ense": -116.618, "epme": -66.985, "epme2": -66.985, "erdek": 27.845, "erdem": 34.328, "espe": 121.895, "estr": -71.314, "euro": 3.275, "exmt": 3.424, "faja": -65.631, "fbfl": -81.465, "fbfl2": -81.465, "fer1": -8.326, "fer2": -8.249, "ferg": 147.058, "feth": 29.092, "ffcj": 105.669, "figu": 6.934, "figu2": 6.934, "fish": -4.98, "fmfl": -81.87, "fmfl2": -81.87, "fong": 179.195, "form": 1.419, "fort": -38.467, "fosm": 4.893, "fosm2": 4.893, "fpga": -80.902, "fpga2": -80.902, "fpnt": -122.465, "fpnt2": -122.465, "fren": -166.283, "frtr": -94.27, "ftfr": -61.063, "ftfr2": -61.063, "fue2": -13.857999999999999, "fuer2": -13.856, "fuka": 139.93, "futu": -178.16, "GA37": 13.59, "gamb": -134.969, "gand": -0.151, "ganm": 73.152, "garc": 72.39399999999999, "gbit": 175.489, "gcsb": 153.433, "GE25": 8.925, "geor": -81.383, "GI20": 15.193, "gibr": -5.35, "gibr2": -5.35, "gibr3": -5.35, "gij2": -5.6979999999999995, "gila": -89.95700000000001, "gist": 178.023, "gokc": 25.894000000000002, "goto": 11.78, "goug": -9.88, "gpab": 114.565, "gptx": -94.79299999999999, "greg": -70.209, "greg2": -70.209, "groo": 136.416, "guam": 144.65200000000002, "gvd9": 24.12, "gwda": 62.339, "hade": 34.863, "hade2": 34.883, "haif": 35.007, "hain": 94.318, "hako": 140.72, "hali": -63.58, "hamb": 81.127, "hana": 145.57, "hani": 73.167, "harl": 5.4110000000000005, "harv": -120.667, "harw": 1.29, "hasp": 0.573, "hbay": 1.115, "helg": 7.891, "hens": -133.0, "herb": -2.299, "herb2": -2.299, "heys": -2.92, "hie2": -17.902, "hien": 164.942, "hill": 115.73899999999999, "hilo": -155.058, "hilo2": -155.058, "hink": -3.13, "hirt": 9.97, "hiva": -139.034, "hmda": 132.07, "hoek": 4.119, "holy": -4.6339999999999995, "holy2": -4.62, "honn": 25.983, "hono": -157.86700000000002, "hono2": -157.86700000000002, "horn": 8.297, "hrak": 25.15, "htst": -157.816, "huahi": -151.032, "huas2": -71.25, "huas3": -71.25, "huat": -96.117, "huat2": -96.117, "huel": -6.834, "iaix": -1.1740000000000002, "ibiz": 1.45, "iclr": -114.723, "iera": 25.74, "igne": 28.024, "iler": -52.586999999999996, "iler2": -52.586999999999996, "ilfa": -4.11, "IM01": 8.019, "im02": 8.019, "im02a": 8.019, "imbi": -48.651, "imbt": -48.651, "immi": -0.19, "imuj": -86.74600000000001, "imuj2": -86.74600000000001, "inav": -75.75, "inav2": -75.75, "inha": 35.5, "iqui": -70.148, "iqui2": -70.148, "isab": -65.435, "ishig": 124.16, "iske": 36.177, "ista": 29.074, "IT45": 15.502, "itea": 22.42, "jaca": -72.535, "jack": 168.61599999999999, "jask": 57.77, "jask2": 57.77, "jers": -2.12, "john": -169.523, "jrmi": -74.11, "juan": -78.83, "juan2": -78.83, "kaci": 34.05, "kahu": -156.472, "kahu2": -156.472, "kait": 173.703, "kala": 22.11, "kalt": 28.07, "kant": -171.718, "kapi": 154.80700000000002, "kaps": 23.0, "kara": 66.975, "kast": 23.636999999999997, "kata": 21.319000000000003, "kaum": -156.9, "kawa": -155.832, "kawa2": -155.832, "keba": 67.43, "kepo": -36.5, "kepo1": -36.5, "kerg": 70.22, "kerg2": 70.02199999999999, "khol": 142.04399999999998, "kiel": 10.275, "kinl": -5.05, "kis1": -5.922000000000001, "kjni": 167.954, "kodi": -152.512, "kodi2": -152.512, "koli": 106.891, "koli2": 106.891, "koro": 21.96, "kors": 142.767, "kos1": 27.288, "kos2": 27.304000000000002, "kota": 98.417, "kril": 142.08, "krna": 81.32, "kuri": 147.88, "kush": 144.37, "kusm": 135.77, "kwaj": 167.738, "kwaj2": 167.738, "kwfl": -81.808, "kwfl2": -81.808, "kypa": 21.66, "LA23": 12.604000000000001, "LA38": 9.857999999999999, "lago": -17.108, "lagos": 3.407, "lajo": -117.258, "lajo2": -117.258, "lalb": -89.319, "lali": -80.906, "lame": -64.72399999999999, "lame2": -64.72399999999999, "lamu": 40.9, "lang": -8.53, "lank": 99.76700000000001, "lapa": -17.768, "larn": 33.641, "larp": -1.2209999999999999, "larp2": -1.2209999999999999, "laru2": 55.5, "lasp": -15.412, "lata": 165.80200000000002, "laun": -87.818, "laza": -102.178, "lcst": 17.029, "lebu": -73.664, "lebu2": -73.664, "leco": -4.78, "leco2": -4.78, "lecy": -2.895, "lega": 123.758, "leha": 0.106, "leha2": 0.106, "leit": -3.18, "lemba": 116.072, "lena": 169.266, "lero": -60.938, "lerw": -1.145, "lerw2": -1.145, "leso": -1.794, "leso2": -1.794, "levu": 177.438, "lgos": -8.667, "LI11": 10.299000000000001, "lifo": 167.27900000000002, "lime": -64.75399999999999, "lime2": -64.75399999999999, "limon": -83.02, "lirf": 152.145, "litz": 167.44400000000002, "live": -3.02, "llan": -3.83, "lmma": 4.308, "lomb": 147.374, "losa": -118.272, "losa2": -118.272, "losu": -84.666, "lott": 178.15900000000002, "lowe": 1.75, "lpaz": -110.333, "lpaz2": -110.333, "lpbc": -133.1, "lton": -111.30799999999999, "luba": 120.20200000000001, "luga": 167.18900000000002, "lymg": -1.507, "made": -92.40100000000001, "madry": -65.031, "maer": 27.967, "magi": -67.046, "magi2": -67.046, "maho": 4.271, "mais": 37.867, "mais2": 37.867, "maji": 56.606, "maju": 105.84100000000001, "maka": -157.66899999999998, "make": -143.569, "mal3": -4.417, "mala": 134.45, "male": 73.527, "malh": -7.334, "mali": -7.334, "malk": 146.82, "malo": 5.117, "malp": -81.604, "malp2": -81.604, "malt": 14.332, "mang": 28.595, "mani": 120.97200000000001, "manz": -104.29799999999999, "mare": 167.877, "mari": -8.691, "marm": 73.8, "mars": 5.356, "marsh": 171.373, "masi": 58.867, "mata": -72.109, "matel": 8.72, "maya": -67.15899999999999, "maza": -106.42399999999999, "maza2": -106.42399999999999, "MC41": 10.238, "ME13": 15.562999999999999, "meji": -70.45100000000001, "meji2": -70.45100000000001, "meli": -2.929, "ment": 26.717, "mera": 139.83, "meul": 96.132, "meul2": 96.132, "mhav": -5.05, "midx": -177.356, "midx2": -177.356, "mill": -4.91, "mimz": -1.295, "mimz2": -1.295, "mind": -25.000999999999998, "mini": 73.05, "mins": 153.98, "mnkt": 174.512, "mnza": -104.301, "moku": -157.793, "momb": 39.65, "mona": -67.938, "mona2": -67.938, "monc": 7.422000000000001, "monc2": 7.422000000000001, "mont": -121.887, "mont2": -121.887, "mony": -71.96, "mony2": -71.96, "motr": -3.5239999999999996, "moul": 97.617, "mrig": -61.283, "mrms": 28.385, "ms001": 99.734, "ms002": 100.182, "ms003": 98.939, "ms004": 102.76700000000001, "ms005": 116.846, "ms006": 119.07799999999999, "mtwa": 40.2, "mumb": -3.98, "murc2": -0.9740000000000001, "mus2": 58.565, "musc": 58.567, "NA23": 14.269, "naga": 129.87, "naha": 127.67, "naho": 132.86, "nain": -61.68, "nanc": 93.55, "nans": 112.5, "napt": 176.92, "nauu": 166.90900000000002, "nawi": -159.36, "nawi2": -159.36, "ncpt": 173.05, "neve": 141.86, "newl": -5.542999999999999, "newl2": -5.542999999999999, "nhav": 0.06, "nhte": -73.431, "nice": 7.285, "nice2": 7.285, "niue": -169.921, "nkfa": -175.18099999999998, "nkfa2": -175.18099999999998, "nksk": 166.02, "noct": -15.95, "nome": -165.43, "nome2": -165.43, "noto": 137.16, "noua": -17.035999999999998, "npor": -2.99, "nshi": -1.44, "nspi": -124.21700000000001, "nspi2": -124.21700000000001, "ntue": -73.431, "ntue2": -73.431, "nuku": -140.085, "numbo": 166.412, "nuuk": -51.72, "nyal": 11.95, "ofun": 141.75, "ohig": -57.901, "ohig3": -57.901, "omae": 138.22, "OR24": 14.415, "oran": -70.033, "orma": 64.676, "oste": 2.9210000000000003, "OT15": 18.497, "otat": 170.62900000000002, "ouin": 166.683, "ouis": -0.249, "ouis2": -0.249, "ouve": 166.562, "PA07": 13.370999999999999, "pada": 100.367, "padn": 100.367, "pagb": 144.797, "pagi": -73.527, "pagi2": -73.527, "pago": -170.69099999999997, "pago2": -170.69099999999997, "pagx": -170.69099999999997, "palb": -124.814, "pale": 23.68, "palm": 2.637, "palm1": -22.983, "pana": 108.501, "pand": 108.501, "pano": 22.254, "pant": 11.937000000000001, "papa": -85.656, "pape": -149.57299999999998, "pape2": -149.57299999999998, "papho": 32.409, "papo": -70.469, "papo2": -70.469, "para": 34.037, "parh": -61.783, "pata": -70.19800000000001, "pata2": -70.19800000000001, "pbil": -83.383, "pblu": -83.69200000000001, "pcha": -72.82, "pcha2": -72.82, "pcoi": -83.068, "pcor": -87.959, "pcrt": -87.167, "pdas": -25.68, "PE09": 13.527000000000001, "pedn": -74.40899999999999, "pedn2": -74.40899999999999, "peir": 23.621, "pemba": 40.483000000000004, "penr": -158.05100000000002, "petr": 158.65, "phbc": -127.48299999999999, "phcp": 118.42, "pich": -71.529, "pich2": -71.529, "pich3": -71.529, "pisa": -70.21600000000001, "pisa2": -70.21600000000001, "pkem": 150.912, "PL14": 15.275, "plat": -57.55, "plat2": -57.538999999999994, "plmy": -162.09, "plom": 26.371, "pluz": -15.41, "plym": -4.19, "pmel": -73.748, "pmel2": -73.748, "pmon": -72.961, "pmon2": -72.961, "pmur": 114.191, "pnfl": -87.212, "pnfl2": -87.212, "pntn": 11.833, "PO40": 12.966, "pobe": -88.20100000000001, "porl": 141.613, "poro": 143.09, "porp": -5.12, "pors": 138.467, "port": -1.062, "port2": -1.062, "posa": -16.315, "posi": 130.8, "pots": -87.50399999999999, "ppcp": 15.123, "prai": -23.502, "prat": -59.663000000000004, "prat3": -59.663000000000004, "prba": -88.62200000000001, "prec": -61.227, "preo": 133.922, "prev": 20.729, "prgi": 111.73299999999999, "pric": -61.765, "prig": 111.73299999999999, "prin": -130.32, "prin2": -130.317, "prog": -89.667, "prog2": -89.667, "prte": 25.633000000000003, "prud": -148.51, "prud2": -148.51, "prus": -6.66, "psan": -86.765, "psdn": -86.764, "psjs": -85.875, "psla": -89.40700000000001, "psla2": -89.40700000000001, "pslu": -120.75299999999999, "pslu2": -120.75299999999999, "PT17": 8.404, "ptal": -71.608, "ptal2": -71.608, "ptan": -96.492, "ptan2": -96.492, "ptar": -70.862, "ptar2": -70.862, "ptbl": 92.76, "ptca": -68.376, "ptch": -71.469, "ptfe": 6.718, "ptfe2": 6.718, "ptga": -60.992, "ptis": -4.834, "ptln": 3.037, "ptln2": 3.037, "ptlu": 57.504, "ptma": 14.494000000000002, "ptmd": -61.468, "ptmd2": -61.458, "ptme": -70.247, "ptmt": -1.11, "ptos": -51.168, "ptpl": -70.702, "ptpr": -72.38, "ptpt": -61.531000000000006, "ptpt2": -61.531000000000006, "ptro": -76.846, "ptsc": -59.645, "ptsp": -61.516999999999996, "ptve": 3.1069999999999998, "ptve2": 3.1069999999999998, "puer2": -105.243, "puert": -105.243, "pumo": -86.867, "pumo2": -86.867, "puyt": 166.58900000000003, "pwil": -67.611, "pwil2": -67.611, "qaqo": -46.035, "qcbc": -132.067, "qing": 110.82, "qtro": -71.525, "qtro2": -71.525, "qtro3": -71.525, "quar": 114.213, "quel": -73.215, "quel2": -73.215, "quepo": -84.17, "quin": 109.25399999999999, "quir": -73.057, "quir2": -73.057, "qura": 58.925, "RA10": 12.283, "rangi": -147.70600000000002, "raro": -159.783, "raro2": -159.785, "rbct": -177.894, "RC09": 15.649000000000001, "reun": 55.28, "reun2": 55.28, "reyk": -21.933000000000003, "rfrt": -177.90400000000002, "riki": -134.967, "rodr": 63.42100000000001, "rorv": 11.25, "rosc": -3.966, "rosc2": -3.966, "rose": -61.388999999999996, "ross": 150.79, "rote": 123.06, "rothe": -68.133, "rous": 8.937999999999999, "rous2": 8.937999999999999, "rtas": -86.54, "rudn": 135.8, "SA16": 14.751, "sab2": 95.333, "saba": 95.333, "sade": 110.79799999999999, "sado": 138.52, "sadp": 110.79899999999999, "sagr": -8.927999999999999, "sagu": -0.20600000000000002, "said": -2.293, "saig": 133.33, "sain": -2.2, "sain2": -2.2, "saip": 145.74200000000002, "sala": 54.007, "sali": -95.197, "sali2": -95.197, "salv": -38.516999999999996, "Salv2": -38.516999999999996, "sama": -74.222, "sams": 36.336999999999996, "san2": -3.7910000000000004, "sana": -81.70100000000001, "sand": -117.17299999999999, "sanf": -80.109, "sanf3": -80.109, "sanj": -66.117, "sanj2": -66.117, "sano": -71.618, "sano2": -71.618, "sanp": -74.866, "sant": -90.307, "saot": 6.737, "sapz": -77.365, "sapz2": -77.365, "sass": 13.642999999999999, "saum": 131.291, "SB36": 13.89, "sbea": -124.04299999999999, "sbea2": -124.04299999999999, "SC43": 13.075999999999999, "scar": -60.733000000000004, "sche": 4.263, "scoa": -1.682, "scoa2": -1.682, "scor": -21.962, "sdom": -69.62899999999999, "sdpr": -1.153, "sdpt": -160.502, "sdpt2": -160.502, "sebe": 105.512, "sebla": 101.59899999999999, "sema": 110.42, "sete": 3.699, "sete2": 3.699, "setp1": -78.983, "setu": -8.93, "sev2": -6.007999999999999, "seve": 156.139, "seve2": 156.139, "sewa": -149.42700000000002, "sewa2": -149.42700000000002, "shee": 0.74, "shek": 113.89399999999999, "shen": 113.88, "sho2": -71.628, "sian": -87.446, "sib2": 98.8, "sibo": 98.8, "sile": 29.604, "sino": 35.15, "sisa": -90.04799999999999, "sitin": 18.433, "sitk": -135.342, "sitk2": -135.342, "sitt": 92.9, "sjds": -85.875, "slds": -73.618, "smag": -93.855, "smag2": -93.855, "smar": 55.537, "smar2": 55.537, "smog": 11.22, "sobr": 17.62, "sole": 9.374, "sole2": 9.374, "solo": 159.95600000000002, "sona": 9.126, "sosu": 138.33, "sove": 140.29, "spjy": 146.975, "spmi": -56.168, "sprg": 147.933, "sscr": -2.7, "ST44": 15.252, "stan": -57.933, "star": 142.82, "stari": 16.576, "stcr": -64.69800000000001, "stcr2": -64.69800000000001, "sthl": -5.7170000000000005, "sthl2": -5.7139999999999995, "sthm": 18.07, "stjo": -52.72, "stlo2": -16.508, "stlu": -60.997, "stlu2": -60.886, "stlu3": -61.06, "stlu4": -60.953, "stma": -2.028, "stma2": -2.028, "stmr": -6.317, "stmt": -63.085, "stor": -6.39, "stor2": -6.39, "stpa": 77.538, "stqy": -2.822, "stqy2": -2.822, "subi": 120.28299999999999, "sumt": 172.773, "sura": 112.741, "suro": 59.52, "swpr": -1.949, "syow": 39.57, "syro": 24.941, "TA18": 17.224, "tako": -1.745, "tala": -81.28, "talc": -73.10600000000001, "talc2": -73.10600000000001, "talt": -70.49, "talt2": -70.49, "talt3": -70.49, "tanjo": 103.852, "tara": 172.93, "tare": 156.409, "tari": -5.604, "tarr": 1.213, "tasu": 33.836, "taut": 176.18099999999998, "tbwc": 149.92700000000002, "tdcu": -12.3, "tela": -87.45299999999999, "telc": -89.30799999999999, "telu": 97.822, "tema": 0.012, "tene": -16.241, "ters": 5.332999999999999, "tfbc": -125.91, "thes": 22.91, "thev": 133.641, "thio": 166.24200000000002, "thul": -68.863, "tjls": 105.663, "tkdi": -1.745, "tn031": -16.718, "tnbl": 98.5, "tnpr": -3.492, "toam2": 49.428000000000004, "tobe": -6.06, "toco2": -70.212, "toco3": -70.212, "tofi": -125.91, "tort": -64.608, "tosa": 132.96, "toul": 5.912999999999999, "toul2": 5.912999999999999, "toya": 137.22, "TR22": 13.755999999999998, "trab": 39.744, "treg": 7.567, "trin": 81.2, "trnz": 3.82, "trst": 142.222, "tubua": -149.476, "tuca": -71.15, "tudy": -3.447, "tumc": -78.729, "tumc2": -78.729, "tuxp": -97.329, "tuxp2": -97.329, "ugle": 142.07, "ulla": -5.16, "upol": -171.761, "ushu": -68.217, "util": -86.895, "vair": -149.295, "vale": -0.311, "valp": -71.628, "valp2": -71.628, "valp3": -71.628, "vanu": 168.308, "vard": 31.1, "vati": 177.761, "VE19": 12.427, "vela": 16.715, "ver1": -64.27, "vera": -96.124, "vera2": -96.124, "verav": 70.36, "vern": -64.27, "vhbc": -123.117, "VI12": 16.177, "vibc": -123.37, "vieq": -65.471, "vieq2": -65.471, "vig2": -8.726, "vil2": -8.77, "vish": 83.28, "viti": 178.424, "vlad": 131.92, "vlis": 3.597, "vodo": 157.989, "vung": 107.071, "waik": 119.219, "wait": -176.56099999999998, "waka": 141.69, "wake": 166.618, "wake2": 166.618, "wall": -176.101, "walvi": 14.499, "warn": 12.103, "wbhb": -2.764, "weym": -2.45, "whbc": -128.3, "whit": -0.61, "wick": -3.09, "will": -68.942, "wint": -129.03, "wlgt": 174.77900000000002, "wlms": -89.32600000000001, "wlms2": -89.32600000000001, "wood": -70.672, "wood2": -70.672, "work": -3.57, "wpwa": -124.11, "wpwa2": -124.11, "wuda": 57.52, "xish": 112.33, "xmas": -157.47299999999998, "yabu": -65.833, "yaku": -139.735, "yaku2": -139.735, "yalo": 29.278000000000002, "yapi": 138.125, "yuzh": 145.85, "zanz": 39.183, "zhap": 111.82, "zihu": -101.55799999999999, "zihu2": -101.55799999999999, "zygi": 33.338, "zygi1": 33.34}
+//go:embed sea_stations_pos.json
+var stationsPosJson []byte
+
+var stationsPos map[string]struct {
+	Lat float64 `json:"lat"`
+	Lon float64 `json:"lon"`
+}
+
+func init() {
+	if err := json.Unmarshal(stationsPosJson, &stationsPos); err != nil {
+		slog.Error("Failed to unmarshal sea stations's position json", "error", err)
+		os.Exit(1)
+	}
+}
 
 func seaHeight() {
 	resp, err := http.Get("https://www.ioc-sealevelmonitoring.org/list.php?operator=&showall=all&output=general")
 	if err != nil {
-		logger.Info(err.Error())
+		slog.Error("Failed to fetch IOC sea level data", "error", err)
 		return
 	}
 	defer func() { _ = resp.Body.Close() }()
 
 	if resp.StatusCode != http.StatusOK {
-		logger.Error("not 200 code", zap.Int("code", resp.StatusCode))
+		slog.Error("HTTP request failed", "status_code", resp.StatusCode)
 		return
 	}
 	doc, err := goquery.NewDocumentFromReader(resp.Body)
 	if err != nil {
-		logger.Info(err.Error())
+		slog.Error("Failed to parse IOC response", "error", err)
 		return
 	}
 
@@ -46,19 +63,15 @@ func seaHeight() {
 		}
 
 		var code = strings.TrimSpace(td.Eq(0).Text())
-		lat, ok := lats[code]
+		pos, ok := stationsPos[code]
 		if !ok {
 			return
 		}
-		lon, ok := lons[code]
-		if !ok {
-			return
-		}
-		data = append(data, db.StationSeaLevel{Code: code, Lat: lat, Lon: lon, Level: level})
+		data = append(data, db.StationSeaLevel{Code: code, Lat: pos.Lat, Lon: pos.Lon, Level: level})
 	})
 	err = db.UpdateSeaLevel(data)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to update sea level data", "error", err)
 	}
 }
 
@@ -87,7 +100,7 @@ type glossStations struct {
 func stationInfoGlossAll() {
 	resp, err := http.Get("https://www.psmsl.org/products/gloss/data/glossCoreNetwork.xml")
 	if err != nil {
-		logger.Info(err.Error())
+		slog.Error("Failed to fetch GLOSS core network data", "error", err)
 		return
 	}
 	defer func() { _ = resp.Body.Close() }()
@@ -98,7 +111,7 @@ func stationInfoGlossAll() {
 	var tmp glossCoreNetwork
 	err = xml.NewDecoder(resp.Body).Decode(&tmp)
 	if err != nil {
-		logger.Info(err.Error())
+		slog.Error("Failed to decode GLOSS XML data", "error", err)
 		return
 	}
 	tmp.GlossStations.Stations = append(tmp.GlossStations.Stations, db.GlossStation{
@@ -117,7 +130,7 @@ func stationInfoGlossAll() {
 	})
 	err = db.UpdateStationInfoGlossAll(tmp.GlossStations.Stations)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to update GLOSS station data", "error", err)
 		return
 	}
 }
