@@ -2,10 +2,12 @@ package controller
 
 import (
 	"encoding/json"
-	"github.com/google/uuid"
-	"go.uber.org/zap"
+	"log/slog"
+
 	"tide/common"
 	"tide/tide_server/db"
+
+	"github.com/google/uuid"
 )
 
 func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk bool) {
@@ -20,7 +22,7 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 	)
 
 	if ds, err := db.GetDevices(stationId); err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to get devices for station", "station_id", stationId, "error", err)
 		return
 	} else {
 		for _, device := range ds {
@@ -29,7 +31,7 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 	}
 
 	if items, err := db.GetItems(stationId); err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to get items for station", "station_id", stationId, "error", err)
 		return
 	} else {
 		for _, item := range items {
@@ -42,7 +44,7 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 			// only create device
 			var device = db.Device{StationId: stationId, Name: deviceName}
 			if err = db.EditDevice(device); err != nil { // add device
-				logger.Error(err.Error())
+				slog.Error("Failed to add device", "station_id", stationId, "device_name", deviceName, "error", err)
 				return
 			}
 			Publish(configPubSub, SendMsgStruct{Type: kMsgSyncDevice, Body: device}, nil)
@@ -57,7 +59,7 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 	for deviceName, keep := range oldDevices {
 		if !keep {
 			if _, err = db.DelDevice(stationId, deviceName); err != nil {
-				logger.Error(err.Error())
+				slog.Error("Failed to delete device", "station_id", stationId, "device_name", deviceName, "error", err)
 				return
 			}
 			Publish(configPubSub, SendMsgStruct{Type: kMsgDelDevice, Body: db.Device{StationId: stationId, Name: deviceName}}, nil)
@@ -68,7 +70,7 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 		if _, ok := newItems[itemName]; !ok {
 			// delete item
 			if _, err = db.DelItem(stationId, itemName); err != nil {
-				logger.Error(err.Error())
+				slog.Error("Failed to delete item", "station_id", stationId, "item_name", itemName, "error", err)
 				return
 			}
 			Publish(configPubSub, SendMsgStruct{Type: kMsgDelItem, Body: db.Item{StationId: stationId, Name: itemName}}, nil)
@@ -84,7 +86,7 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 		if !ok || oldItemInfo.Name != newItemInfo.DeviceName || oldItemInfo.Type != newItemInfo.Type {
 			item := db.Item{StationId: stationId, Name: itemName, Type: newItemInfo.Type, DeviceName: newItemInfo.DeviceName}
 			if err = db.EditItem(item); err != nil {
-				logger.Error(err.Error())
+				slog.Error("Failed to edit item", "station_id", stationId, "item_name", itemName, "error", err)
 				return
 			}
 			Publish(configPubSub, SendMsgStruct{Type: kMsgSyncItem, Body: item}, nil)
@@ -95,11 +97,11 @@ func SyncStationInfo(stationId uuid.UUID, info common.StationInfoStruct) (retOk 
 	}
 
 	if camerasJson, err := json.Marshal(info.Cameras); err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to marshal cameras JSON", "station_id", stationId, "error", err)
 		return
 	} else {
 		if n, err := db.SyncStationCannotEdit(stationId, camerasJson); err != nil {
-			logger.Error(err.Error())
+			slog.Error("Failed to sync station cameras", "station_id", stationId, "error", err)
 			return
 		} else if n > 0 {
 			Publish(configPubSub, SendMsgStruct{Type: kMsgSyncStationCannotEdit, Body: db.Station{Id: stationId, Cameras: camerasJson}}, nil)
@@ -118,12 +120,12 @@ func WriteItemsLatest(encoder *json.Encoder, stationId uuid.UUID, devices map[st
 	}
 	err := db.GetItemsLatest(stationId, itemsLatest)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to get items latest timestamps", "station_id", stationId, "error", err)
 		return
 	}
-	logger.Debug("get the last record time of each item")
+	slog.Debug("Retrieved latest timestamps for all items", "station_id", stationId, "item_count", len(itemsLatest))
 	if err = encoder.Encode(itemsLatest); err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to encode items latest timestamps", "station_id", stationId, "error", err)
 		return
 	}
 	return true
@@ -133,12 +135,12 @@ func WriteItemsLatest(encoder *json.Encoder, stationId uuid.UUID, devices map[st
 func WriteLatestStatusLogRowId(encoder *json.Encoder, stationId uuid.UUID) (retOk bool) {
 	latestStatusLogRowId, err := db.GetLatestStatusLogRowId(stationId)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to get latest status log row ID", "station_id", stationId, "error", err)
 		return
 	}
-	logger.Debug("latest status log", zap.Int64("rowId", latestStatusLogRowId))
+	slog.Debug("Retrieved latest status log row ID", "station_id", stationId, "row_id", latestStatusLogRowId)
 	if err = encoder.Encode(latestStatusLogRowId); err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to encode latest status log row ID", "station_id", stationId, "error", err)
 		return
 	}
 	return true
@@ -148,14 +150,14 @@ func ReadMissData(decoder *json.Decoder, stationId uuid.UUID) (retOk bool) {
 	var missData map[string][]common.DataTimeStruct
 	err := decoder.Decode(&missData)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to decode miss data", "station_id", stationId, "error", err)
 		return
 	}
 	for itemName, ds := range missData {
-		logger.Debug("got miss data", zap.String("item_name", itemName), zap.Int("len", len(ds)))
+		slog.Debug("Processing miss data", "station_id", stationId, "item_name", itemName, "data_count", len(ds))
 		for _, dataTime := range ds {
 			if n, err := db.SaveDataHistory(stationId, itemName, dataTime.Value, dataTime.Millisecond.ToTime()); err != nil {
-				logger.Error(err.Error())
+				slog.Error("Failed to save miss data history", "station_id", stationId, "item_name", itemName, "error", err)
 				return
 			} else if n > 0 {
 				stationItem := common.StationItemStruct{StationId: stationId, ItemName: itemName}
@@ -165,7 +167,7 @@ func ReadMissData(decoder *json.Decoder, stationId uuid.UUID) (retOk bool) {
 					DataTimeStruct:    dataTime,
 				}, stationItem)
 				if err != nil {
-					logger.DPanic("publish", zap.Error(err))
+					slog.Error("Failed to publish miss data", "station_id", stationId, "item_name", itemName, "error", err)
 				}
 			}
 		}
@@ -177,10 +179,10 @@ func ReadMissStatusLogs(decoder *json.Decoder, stationId uuid.UUID) (retOk bool)
 	var missStatusLogs []common.RowIdItemStatusStruct
 	err := decoder.Decode(&missStatusLogs)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to decode miss status logs", "station_id", stationId, "error", err)
 		return
 	}
-	logger.Debug("got miss logs", zap.Int("len", len(missStatusLogs)))
+	slog.Debug("Processing miss status logs", "station_id", stationId, "log_count", len(missStatusLogs))
 	// The last index of each item
 	var latestLogIndexByItem = make(map[string]int)
 	for i, statusLog := range missStatusLogs {
@@ -190,7 +192,7 @@ func ReadMissStatusLogs(decoder *json.Decoder, stationId uuid.UUID) (retOk bool)
 		statusLog := missStatusLogs[i]
 		_, err = db.UpdateItemStatus(stationId, statusLog.ItemName, statusLog.Status, statusLog.ChangedAt.ToTime())
 		if err != nil {
-			logger.Error(err.Error())
+			slog.Error("Failed to update item status from miss logs", "station_id", stationId, "item_name", statusLog.ItemName, "error", err)
 			return
 		}
 	}
@@ -198,7 +200,7 @@ func ReadMissStatusLogs(decoder *json.Decoder, stationId uuid.UUID) (retOk bool)
 	for _, statusLog := range missStatusLogs {
 		n, err := db.SaveItemStatusLog(stationId, statusLog.RowId, statusLog.ItemName, statusLog.Status, statusLog.ChangedAt.ToTime())
 		if err != nil {
-			logger.Error(err.Error())
+			slog.Error("Failed to save item status log", "station_id", stationId, "item_name", statusLog.ItemName, "row_id", statusLog.RowId, "error", err)
 			return
 		} else if n > 0 {
 			err = configPubSub.Publish(SendMsgStruct{Type: kMsgMissItemStatus,
@@ -207,7 +209,7 @@ func ReadMissStatusLogs(decoder *json.Decoder, stationId uuid.UUID) (retOk bool)
 					RowIdItemStatusStruct: statusLog,
 				}}, nil)
 			if err != nil {
-				logger.DPanic("publish", zap.Error(err))
+				slog.Error("Failed to publish miss item status", "station_id", stationId, "item_name", statusLog.ItemName, "error", err)
 			}
 		}
 	}
