@@ -4,13 +4,15 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"github.com/google/uuid"
-	"go.uber.org/zap"
 	"io"
+	"log/slog"
 	"net"
 	"strings"
+
 	"tide/common"
 	"tide/tide_server/db"
+
+	"github.com/google/uuid"
 )
 
 func syncDataClient(conn net.Conn) {
@@ -22,7 +24,7 @@ func syncDataClient(conn net.Conn) {
 		var msg forwardDataStruct
 		if err = decoder.Decode(&msg); err != nil {
 			if err != io.EOF && !errors.Is(err, context.Canceled) && !strings.Contains(err.Error(), "use of closed network connection") {
-				logger.Error(err.Error())
+				slog.Error("Failed to decode sync data message", "error", err)
 			}
 			break
 		}
@@ -30,16 +32,16 @@ func syncDataClient(conn net.Conn) {
 		// First time received
 		if _, ok := receivedItems[msg.StationItemStruct]; !ok {
 			receivedItems[msg.StationItemStruct] = struct{}{}
-			logger.Debug("first receive", zap.String("station_id", msg.StationId.String()), zap.String("item_name", msg.ItemName))
+			slog.Debug("First time receiving data from item", "station_id", msg.StationId, "item_name", msg.ItemName)
 			// First make sure the table exists
 			if err = db.MakeSureTableExist(msg.ItemName); err != nil {
-				logger.Error(err.Error())
+				slog.Error("Failed to create table for item", "item_name", msg.ItemName, "error", err)
 				return
 			}
 		}
 		// save and publish
 		if n, err := db.SaveDataHistory(msg.StationId, msg.ItemName, msg.Value, msg.Millisecond.ToTime()); err != nil {
-			logger.Error(err.Error())
+			slog.Error("Failed to save data history", "station_id", msg.StationId, "item_name", msg.ItemName, "error", err)
 			return
 		} else if n > 0 {
 			if msg.Type == kMsgDataGpio {
@@ -62,7 +64,7 @@ func fillMissDataClient(conn net.Conn) (retOk bool) {
 	var permissions common.UUIDStringsMap
 	err := decoder.Decode(&permissions)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to decode permissions", "error", err)
 		return
 	}
 	var stationsItemsLatest = make(map[uuid.UUID]common.StringMsecMap)
@@ -73,33 +75,33 @@ func fillMissDataClient(conn net.Conn) (retOk bool) {
 		}
 		err = db.GetItemsLatest(stationId, itemsLatest)
 		if err != nil {
-			logger.Error(err.Error())
+			slog.Error("Failed to get items latest timestamps", "station_id", stationId, "error", err)
 			return
 		}
 		stationsItemsLatest[stationId] = itemsLatest
 	}
 
 	if err = encoder.Encode(stationsItemsLatest); err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to encode stations items latest", "error", err)
 		return
 	}
 
 	var stationsMissData map[uuid.UUID]map[string][]common.DataTimeStruct
 	err = decoder.Decode(&stationsMissData)
 	if err != nil {
-		logger.Error(err.Error())
+		slog.Error("Failed to decode stations miss data", "error", err)
 		return
 	}
 	for stationId, missData := range stationsMissData {
 		for itemName, ds := range missData {
-			logger.Debug("got miss data", zap.String("item_name", itemName), zap.Int("len", len(ds)))
+			slog.Debug("Processing miss data", "item_name", itemName, "data_count", len(ds))
 			for _, data := range ds {
 				if err = db.MakeSureTableExist(itemName); err != nil {
-					logger.Error(err.Error())
+					slog.Error("Failed to create table for item", "item_name", itemName, "error", err)
 					return
 				}
 				if n, err := db.SaveDataHistory(stationId, itemName, data.Value, data.Millisecond.ToTime()); err != nil {
-					logger.Error(err.Error())
+					slog.Error("Failed to save miss data history", "station_id", stationId, "item_name", itemName, "error", err)
 					return
 				} else if n > 0 {
 					stationItem := common.StationItemStruct{StationId: stationId, ItemName: itemName}
@@ -109,7 +111,7 @@ func fillMissDataClient(conn net.Conn) (retOk bool) {
 						DataTimeStruct:    data,
 					}, stationItem)
 					if err != nil {
-						logger.DPanic("publish", zap.Error(err))
+						slog.Error("Failed to publish miss data", "station_id", stationId, "item_name", itemName, "error", err)
 					}
 				}
 			}
